Restructured code to make it more reusable
